+++ 22_simulate_incident_robotshop.sh
@@ -1,1 +1,0 @@
-./tools/01_demo/robotshop_incident_memory.sh