+++ tools/05_topology/aimanager/create-merge-rules.sh
@@ -1,6 +1,5 @@
 echo "Create Rules - Starting..."
 export WAIOPS_NAMESPACE=$(oc get po -A|grep aiops-orchestrator-controller |awk '{print$1}')
-export EVTMGR_NAMESPACE=$(oc get po -A|grep noi-operator |awk '{print$1}')