+++ tools/11_practicum/stream_remote_logs.sh
@@ -1,3 +1,3 @@
 
-INSTALL_POD=$(oc get po -n default|grep install|awk '{print$1}')
-oc logs -n default -f $INSTALL_POD
+INSTALL_POD=$(oc get po -n cp4waiops-installation|grep install|awk '{print$1}')
+oc logs -n cp4waiops-installation -f $INSTALL_POD