+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [41.0.1] - 2023-04-24
+
+
+### Major changes
+
+- Deprecation of EventManager
+- Deprecation of AIManager
+- Complete overhaul of file structure
+- Complete overhaul of naming conventions (AIManager-->CP4WAIOPS)
+
+
+
+
 ## [37.0.1] - 2023-04-14